requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-simpleui==2023.12.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
feedparser==6.0.11
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
pillow==10.2.0
prettytable==3.10.0
pyecharts==2.0.5
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
sgmllib3k==1.0.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
